+++ app/javascript/components/drawer.vue
@@ -1,8 +1,7 @@
 <template>
   <v-navigation-drawer app
                        v-model="drawer"
-                       absolute
-                       temporary>
+                       absolute>
     <v-toolbar flat class="transparent">
       <v-list>
         <v-list-tile avatar>
@@ -137,7 +136,7 @@
         },
 
         set(value) {
-          store.commit('setDrawer', value);
+          store.commit('SET_DRAWER', value);
         }
       }
     },