+++ app/javascript/components/reset_password.vue
@@ -82,7 +82,7 @@
       ]),
 
       ...mapMutations([
-        'setAlert'
+        'SET_ALERT'
       ]),
 
       submit() {
@@ -99,7 +99,7 @@
             let type = "success";
             let message = "Password was successful reset.";
 
-            this.setAlert({ type, message });
+            this.SET_ALERT({ type, message });
 
             this.$router.push('/profile');
           } else if (response.response && response.response.status === 422) {