+++ app/javascript/components/sign_in.vue
@@ -92,7 +92,7 @@
       ]),
 
       ...mapMutations([
-        'setAlert'
+        'SET_ALERT'
       ]),
 
       submit() {
@@ -108,7 +108,7 @@
             let type = "success";
             let message = "Successful signed in!";
 
-            this.setAlert({ type, message });
+            this.SET_ALERT({ type, message });
 
             this.$router.push('/profile');
           } else if (response.response && response.response.status === 422) {