+++ app/javascript/components/sign_up.vue
@@ -93,7 +93,7 @@
       ]),
 
       ...mapMutations([
-        'setAlert'
+        'SET_ALERT'
       ]),
 
       submit () {
@@ -110,7 +110,7 @@
             let type = "success";
             let message = "Successful signed up!";
 
-            this.setAlert({ type, message });
+            this.SET_ALERT({ type, message });
 
             this.$router.push('/profile');
           } else if (response.response && response.response.status === 422) {