+++ app/javascript/packs/evemonk_vue.js
@@ -28,6 +28,4 @@ document.addEventListener('DOMContentLoaded', () => {
     store,
     render: h => h(App),
   });
-
-  console.log(app);
 });